+++ lib/byebug/processors/command_processor.rb
@@ -116,7 +116,6 @@ module Byebug
 
     def after_repl
       interface.autosave
-      interface.close
     end
 
     #