+++ lib/byebug/processors/script_processor.rb
@@ -23,6 +23,12 @@ module Byebug
       end
     end
 
+    def after_repl
+      super
+
+      interface.close
+    end
+
     #
     # Prompt shown before reading a command.
     #