+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Show valid breakpoint locations when invalid location given (#393, @ko1).
 * Ruby 2.5.0 support (#397, @yui-knk).
+* Some love & tests to remote debugging (#82).
 
 ## 9.1.0 - 2017-08-22